Bug 739096 - Update cairo to commit 277a1daec80cb6cf7bfb0e200cf78e7842cb2f82 (release 1.17.4 + post-release fixes on trunk) from https://gitlab.freedesktop.org/cairo/cairo/-/tree/master. r=jrmuizel
Differential Revision: https://phabricator.services.mozilla.com/D112558

UltraBlame original commit: d41a2bcb4ff2a9b5ae76160037d06671d103adba

+++ gfx/cairo/cairo/src/cairo-cache-private.h
@@ -84,7 +84,7 @@
 
 
 typedef struct _cairo_cache_entry {
-    uintptr_t hash;
+    unsigned long hash;
     unsigned long size;
 } cairo_cache_entry_t;
